requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.2
cryptography==41.0.4
Flask==3.0.0
Flask-API.yandex==0.6.2.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.3
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pytest==7.4.2
pytest-mock==3.11.1
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
soupsieve==2.5
types-PyYAML==6.0.12.12
urllib3==2.0.5
validators==0.22.0
Werkzeug==3.0.0
